refactor: export ApplyPatch
